+++ src/tasks/bankImage.ts
@@ -8,7 +8,7 @@ import * as path from 'path';
 
 import { allCollectionLogItems } from '../lib/collectionLog';
 import { Events } from '../lib/constants';
-import { filterableTypes } from '../lib/filterables';
+import { filterableTypes } from '../lib/data/filterables';
 import backgroundImages from '../lib/minions/data/bankBackgrounds';
 import { BankBackground } from '../lib/minions/types';
 import { UserSettings } from '../lib/settings/types/UserSettings';
